+++ Pipfile
@@ -14,6 +14,7 @@ python_version = '3'
 
 nose = "*"
 requests-mock = ">=1.3"
+pipenv = "*"
 
 
 [packages]
@@ -22,4 +23,4 @@ sleekxmpp = ">=1.3"
 click = ">=6"
 requests = ">=2.18"
 pycryptodome = ">=3.4"
-"pycountry-convert" = ">=0.5"
+pycountry-convert = ">=0.5"
