--- sucks
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-
-cd ${DIR}
-pipenv run -- python ./sucks.py "$@"
